extracting volcano code into a separate module (step one)
